android_webview/tools/system_webview_shell/test/data/blink-apis/webmidi/requestmidiaccess-permission-denied-expected.txt

Test navigator.requestMIDIAccess API in WebView
onPermissionRequest: android.webkit.resource.MIDI_SYSEX
request denied
PASS: requestMIDIAccess error, reason: NotAllowedError: The request is not allowed by the user agent or the platform in the current context.
TEST FINISHED
